+++ core/policy.php
@@ -25,17 +25,12 @@
 
 namespace core;
 
-require_once MECCANO_CORE_DIR.'/swconst.php';
-require_once MECCANO_CORE_DIR.'/unifunctions.php';
 require_once MECCANO_CORE_DIR.'/extclass.php';
 
 interface intPolicy {
     function __construct(\mysqli $dbLink);
     public function delPolicy($plugin);
-    public function addPolicyToGroup($id); // old name [addGroup]
-    public function delPolicyFromGroup($id); // old name [delGroup]
     public function setFuncAccess($plugin, $func, $groupId, $access = TRUE); // old name [funcAccess]
-    public function checkFuncAccess($plugin, $func); // old name [checkAccess]
     public function installPolicy(\DOMDocument $policy, $validate = TRUE); // old name [install]
     public function groupPolicyList($plugin, $groupId, $code = MECCANO_DEF_LANG);
     public function getPolicyDescById($id);
@@ -87,64 +82,6 @@ class Policy extends ServiceMethods implements intPolicy {
                 $this->setError(ERROR_NOT_EXECUTED, 'delPolicy: something went wrong -> '.$this->dbLink->error);
                 return FALSE;
             }
-        }
-        return TRUE;
-    }
-    
-    public function addPolicyToGroup($id) {
-        $this->zeroizeError();
-        if (!is_integer($id)) {
-            $this->setError(ERROR_INCORRECT_DATA, 'addPolicyToGroup: id must be integer');
-            return FALSE;
-        }
-        $qIsGroup = $this->dbLink->query("SELECT `g`.`id` FROM `".MECCANO_TPREF."_core_userman_groups` `g` "
-                . "WHERE `g`.`id`=$id "
-                . "AND NOT EXISTS ("
-                . "SELECT `a`.`id` "
-                . "FROM `".MECCANO_TPREF."_core_policy_access` `a` "
-                . "WHERE `a`.`groupid`=$id LIMIT 1) ;");
-        if (!$this->dbLink->affected_rows) {
-            $this->setError(ERROR_ALREADY_EXISTS, 'addPolicyToGroup: defined group is not found or already was added');
-            return FALSE;
-        }
-        $qDbFuncs = $this->dbLink->query("SELECT `id` "
-            . "FROM `".MECCANO_TPREF."_core_policy_summary_list` ;");
-        if ($this->dbLink->errno) {
-            $this->setError(ERROR_NOT_EXECUTED, 'addPolicy: unable to get policies -> '.$this->dbLink->error);
-            return FALSE;
-        }
-        if ($id == 1) {
-            $access = 1;
-        }
-        else {
-            $access = 0;
-        }
-        while (list($row) = $qDbFuncs->fetch_row()) {
-            $this->dbLink->query("INSERT INTO `".MECCANO_TPREF."_core_policy_access` (`groupid`, `funcid`, `access`) "
-                    . "VALUES ($id, $row, $access) ;");
-            if ($this->dbLink->errno) {
-                $this->setError(ERROR_NOT_EXECUTED, 'addPolicy: unable to add policy -> '.$this->dbLink->error);
-                return FALSE;
-            }
-        }
-        return TRUE;
-    }
-    
-    public function delPolicyFromGroup($id) {
-        $this->zeroizeError();
-        if (!is_integer($id)) {
-            $this->setError(ERROR_INCORRECT_DATA, 'delPolicyFromGroup: id must be integer');
-            return FALSE;
-        }
-        $this->dbLink->query("DELETE FROM `".MECCANO_TPREF."_core_policy_access` "
-                . "WHERE `groupid`=$id ;");
-        if ($this->dbLink->errno) {
-            $this->setError(ERROR_NOT_EXECUTED, 'addPolicy: unable to delete policy -> '.$this->dbLink->error);
-            return FALSE;
-        }
-        if (!$this->dbLink->affected_rows) {
-            $this->setError(ERROR_NOT_FOUND, 'delPolicyFromGroup: defined group is not found');
-            return FALSE;
         }
         return TRUE;
     }
@@ -204,47 +141,6 @@ class Policy extends ServiceMethods implements intPolicy {
             return FALSE;
         }
         return TRUE;
-    }
-    
-    public function checkFuncAccess($plugin, $func) {
-        $this->zeroizeError();
-        if (!pregPlugin($plugin) || !pregPlugin($func)) {
-            $this->setError(ERROR_INCORRECT_DATA, 'checkFuncAccess: check incoming parameters');
-            return FALSE;
-        }
-        if (isset($_SESSION[AUTH_USER_ID])) {
-            $qAccess = $this->dbLink->query("SELECT `a`.`access` "
-                    . "FROM `".MECCANO_TPREF."_core_policy_access` `a` "
-                    . "JOIN `".MECCANO_TPREF."_core_policy_summary_list` `s` "
-                    . "ON `a`.`funcid`=`s`.`id` "
-                    . "JOIN `".MECCANO_TPREF."_core_userman_groups` `g` "
-                    . "ON `a`.`groupid`=`g`.`id` "
-                    . "JOIN `".MECCANO_TPREF."_core_userman_users` `u` "
-                    . "ON `g`.`id`=`u`.`groupid` "
-                    . "WHERE `u`.`id`=".$_SESSION[AUTH_USER_ID]." "
-                    . "AND `s`.`name`='$plugin' "
-                    . "AND `s`.`func`='$func' "
-                    . "LIMIT 1 ;");
-        }
-        else {
-            $qAccess = $this->dbLink->query("SELECT `n`.`access` "
-                    . "FROM `".MECCANO_TPREF."_core_policy_nosession` `n` "
-                    . "JOIN `".MECCANO_TPREF."_core_policy_summary_list` `s` "
-                    . "ON `n`.`funcid`=`s`.`id` "
-                    . "WHERE `s`.`name`='$plugin' "
-                    . "AND `s`.`func`='$func' "
-                    . "LIMIT 1 ;");
-        }
-        if ($this->dbLink->errno) {
-            $this->setError(ERROR_NOT_EXECUTED, 'checkFuncAccess: something went wrong -> '.$this->dbLink->error);
-            return FALSE;
-        }
-        if (!$this->dbLink->affected_rows) {
-            $this->setError(ERROR_NOT_FOUND, 'checkFuncAccess: policy is not found');
-            return FALSE;
-        }
-        list($access) = $qAccess->fetch_row();
-        return (int) $access;
     }
     
     public function installPolicy(\DOMDocument $policy, $validate = TRUE) {